+++ packages/browser-wallet/src/popup/page-layouts/MainLayout/i18n/en.ts
@@ -20,7 +20,7 @@ const t = {
         connectAccountsRequest: 'Connect accounts',
         addTokens: 'Add tokens',
         idProof: 'Proof of identity',
-        web3IdProof: 'Proof of identity',
+        web3IdProof: 'Proof of identity request',
         request: 'Signature Request',
         connect: 'New connection',
         allowlistingRequest: 'Allowlisting request',