+++ packages/browser-wallet/src/popup/pages/IdProofRequest/DisplayStatement/DisplayStatement.tsx
@@ -25,7 +25,6 @@ import {
     useStatementHeader,
     useStatementValue,
     useStatementName,
-    isoToCountryName,
 } from './utils';
 
 export type StatementLine = {
@@ -34,11 +33,11 @@ export type StatementLine = {
     isRequirementMet: boolean;
 };
 
-type StatementLineProps = StatementLine;
+type StatementLineProps = StatementLine & ClassName;
 
-export function DisplayStatementLine({ attribute, value, isRequirementMet }: StatementLineProps) {
+export function DisplayStatementLine({ attribute, value, isRequirementMet, className }: StatementLineProps) {
     return (
-        <li className="display-statement__line">
+        <li className={clsx(className, 'display-statement__line')}>
             <div className="display-statement__line-attribute">{attribute}:</div>
             <div className="display-statement__line-value">
                 {value}
@@ -188,7 +187,7 @@ export function DisplayRevealStatement({
     className,
     onInvalid,
 }: DisplayRevealStatementProps) {
-    const { t, i18n } = useTranslation('idProofRequest', { keyPrefix: 'displayStatement' });
+    const { t } = useTranslation('idProofRequest', { keyPrefix: 'displayStatement' });
     const getAttributeName = useGetAttributeName();
     const displayAttribute = useDisplayAttributeValue();
     const attributes =
@@ -197,11 +196,7 @@ export function DisplayRevealStatement({
 
     const lines: StatementLine[] = statements.map((s) => {
         const raw = attributes[s.attributeTag];
-        let value = displayAttribute(s.attributeTag, attributes[s.attributeTag] ?? '');
-
-        if (value && ['countryOfResidence', 'nationality', 'idDocIssuer'].includes(s.attributeTag)) {
-            value = isoToCountryName(i18n.resolvedLanguage)(value);
-        }
+        const value = displayAttribute(s.attributeTag, attributes[s.attributeTag] ?? '');
 
         return {
             attribute: getAttributeName(s.attributeTag),