+++ packages/browser-wallet/src/popup/pages/Web3ProofRequest/AccountStatement.tsx
@@ -1,93 +1,38 @@
 import {
     AccountCredentialStatement,
     createAccountDID,
+    CredentialSchemaSubject,
+    IDENTITY_SUBJECT_SCHEMA,
     RevealStatementV2,
     StatementTypes,
-    AttributeKey,
-    AttributeList,
 } from '@concordium/web-sdk';
-import { displaySplitAddress, useIdentityOf } from '@popup/shared/utils/account-helpers';
-import { useDisplayAttributeValue, useGetAttributeName } from '@popup/shared/utils/identity-helpers';
-import { WalletCredential, ConfirmedIdentity } from '@shared/storage/types';
+import { displaySplitAddress, useIdentityName, useIdentityOf } from '@popup/shared/utils/account-helpers';
+import { useDisplayAttributeValue } from '@popup/shared/utils/identity-helpers';
+import { ConfirmedIdentity, WalletCredential } from '@shared/storage/types';
 import React, { useCallback, useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
-import { ClassName } from 'wallet-common-helpers';
-import { DisplayStatementView, StatementLine } from '../IdProofRequest/DisplayStatement/DisplayStatement';
+import { SecretStatement, useStatementName, useStatementValue } from '../IdProofRequest/DisplayStatement/utils';
 import CredentialSelector from './CredentialSelector';
+import { DisplayRevealStatements } from './Display/DisplayRevealStatements';
+import { DisplaySecretStatements } from './Display/DisplaySecretStatements';
+import { OverwriteSecretLine } from './Display/utils';
 import { DisplayCredentialStatementProps, SecretStatementV2 } from './utils';
-import {
-    isoToCountryName,
-    SecretStatement,
-    useStatementDescription,
-    useStatementHeader,
-    useStatementName,
-    useStatementValue,
-} from '../IdProofRequest/DisplayStatement/utils';
 
-type DisplaySecretStatementV2Props = ClassName & {
-    identity?: ConfirmedIdentity;
-    dappName: string;
-    statement: SecretStatementV2;
-};
+export function DisplayAccount({ option }: { option: WalletCredential }) {
+    const identityName = useIdentityName(option);
 
-export function DisplaySecretStatementV2({ dappName, statement, identity, className }: DisplaySecretStatementV2Props) {
-    const v1Statement: SecretStatement = statement as SecretStatement;
-    const header = useStatementHeader(v1Statement);
-    const value = useStatementValue(v1Statement);
-    const description = useStatementDescription(v1Statement, identity);
-    const attribute = useStatementName(v1Statement);
-
-    const lines: StatementLine[] = [
-        {
-            attribute,
-            value,
-            isRequirementMet: identity !== undefined,
-        },
-    ];
+    if (!identityName) {
+        return null;
+    }
 
     return (
-        <DisplayStatementView
-            lines={lines}
-            dappName={dappName}
-            header={header}
-            description={description}
-            className={className}
-        />
+        <header className="verifiable-credential__header">
+            <div className="web3-id-proof-request__selector-title flex-column align-start">
+                <div className="display5">{displaySplitAddress(option.address)}</div>
+                <div className="bodyS">{identityName}</div>
+            </div>
+        </header>
     );
-}
-
-type DisplayRevealStatementV2Props = ClassName & {
-    identity?: ConfirmedIdentity;
-    dappName: string;
-    statements: RevealStatementV2[];
-};
-
-export function DisplayRevealStatementV2({ dappName, statements, identity, className }: DisplayRevealStatementV2Props) {
-    const { t, i18n } = useTranslation('idProofRequest', { keyPrefix: 'displayStatement' });
-    const getAttributeName = useGetAttributeName();
-    const displayAttribute = useDisplayAttributeValue();
-    const header = t('headers.reveal');
-    const attributes = identity
-        ? identity.idObject.value.attributeList.chosenAttributes
-        : ({} as AttributeList['chosenAttributes']);
-
-    const lines: StatementLine[] = statements.map((s) => {
-        const stringTag = s.attributeTag as AttributeKey;
-        const raw = attributes[stringTag];
-        let value = displayAttribute(stringTag, raw ?? '');
-
-        if (value && ['countryOfResidence', 'nationality', 'idDocIssuer'].includes(stringTag)) {
-            value = isoToCountryName(i18n.resolvedLanguage)(value);
-        }
-
-        return {
-            attribute: getAttributeName(stringTag),
-            value: value ?? 'Unavailable',
-            isRequirementMet: raw !== undefined,
-        };
-    });
-
-    return <DisplayStatementView reveal lines={lines} dappName={dappName} header={header} className={className} />;
 }
 
 export default function AccountStatement({
@@ -96,13 +41,16 @@ export default function AccountStatement({
     dappName,
     setChosenId,
     net,
+    showDescription,
 }: DisplayCredentialStatementProps<AccountCredentialStatement, WalletCredential>) {
+    const { t } = useTranslation('web3IdProofRequest');
     const reveals = credentialStatement.statement.filter(
         (s) => s.type === StatementTypes.RevealAttribute
     ) as RevealStatementV2[];
     const secrets = credentialStatement.statement.filter(
         (s) => s.type !== StatementTypes.RevealAttribute
     ) as SecretStatementV2[];
+    const displayAttribute = useDisplayAttributeValue();
 
     const [chosenCredential, setChosenCredential] = useState<WalletCredential | undefined>(validCredentials[0]);
     // We do the type cast, because the check should have been done to filter validCredentials.
@@ -120,31 +68,50 @@ export default function AccountStatement({
         }
     }, []);
 
+    const accountCreateSecretLine: OverwriteSecretLine = (statement: SecretStatementV2) => {
+        const value = useStatementValue(statement as SecretStatement);
+        const attribute = useStatementName(statement as SecretStatement);
+        return {
+            attribute,
+            value,
+        };
+    };
+
+    if (!identity) {
+        return null;
+    }
+
     return (
         <div className="web3-id-proof-request__credential-statement-container">
-            <CredentialSelector
+            {showDescription && (
+                <p className="web3-id-proof-request__description bodyM">{t('descriptions.accountCredential')}</p>
+            )}
+            <CredentialSelector<WalletCredential>
                 options={validCredentials}
-                displayOption={(option) => displaySplitAddress(option.address)}
+                DisplayOption={DisplayAccount}
                 onChange={onChange}
+                header={t('select.accountCredential')}
             />
             {reveals.length !== 0 && (
-                <DisplayRevealStatementV2
-                    className="m-t-10:not-first"
+                <DisplayRevealStatements
+                    className="web3-id-proof-request__statement"
                     dappName={dappName}
-                    identity={identity}
+                    attributes={identity.idObject.value.attributeList.chosenAttributes}
                     statements={reveals}
+                    formatAttribute={displayAttribute}
+                    schema={IDENTITY_SUBJECT_SCHEMA as CredentialSchemaSubject}
                 />
             )}
-            {secrets.map((s, i) => (
-                <DisplaySecretStatementV2
-                    // eslint-disable-next-line react/no-array-index-key
-                    key={i} // Allow this, as we don't expect these to ever change.
-                    className="m-t-10:not-first"
-                    dappName={dappName}
-                    identity={identity}
-                    statement={s}
+            {secrets.length !== 0 && (
+                <DisplaySecretStatements
+                    className="web3-id-proof-request__statement"
+                    attributes={identity.idObject.value.attributeList.chosenAttributes}
+                    statements={secrets}
+                    formatAttribute={displayAttribute}
+                    schema={IDENTITY_SUBJECT_SCHEMA as CredentialSchemaSubject}
+                    overwriteSecretLine={accountCreateSecretLine}
                 />
-            ))}
+            )}
         </div>
     );
 }