+++ packages/browser-wallet/src/popup/pages/Web3ProofRequest/i18n/en.ts
@@ -1,7 +1,8 @@
 export default {
     header: '{{dappName}} requests the following information about you:',
-    accept: 'Accept',
+    approve: 'Approve',
     continue: 'Continue',
+    back: 'Back',
     reject: 'Reject',
     displayStatement: {
         requirementsMet: 'You meet this requirement',
@@ -9,12 +10,12 @@ export default {
         revealDescription:
             '<1>Important:</1> {{dappName}} will be given all the information above. You should only accept, if you trust the service, and you are familiar with their privacy policy.',
         revealTooltip: {
-            header: 'Revealing information <1 />',
-            body: 'When you reveal information to a third party, you effectively hand over the information to them. This means you should only do this if you agree to their data usage and protection policies.\n\nYou can read more in\n<1>the developer documentation</1>.',
+            header: 'Information to reveal',
+            body: 'When you reveal information for a third party, you effectively hand over the information to them. This means that you should only do this if you have absolute trust in them, and if you are familiar with their data usage and protection procedures.\n\nYou can read more on\ndeveloper.concordium.software',
         },
         secretTooltip: {
-            header: 'Zero Knowledge proofs',
-            body: 'Zero Knowledge proofs are a way of proving something to a service or dApp without revealing the exact personal information. One example can be that you prove that you are over 18 years old without revealing your exact age. Another example could be proving your residency is within a given set of countries without revealing which of those countries you reside within.\n\nYou can read more in\n<1>the developer documentation</1>.',
+            header: 'Zero Knowledge proof',
+            body: 'Zero-knowledge proofs are a way of proving something to a service or dApp without revealing the exact personal information. One example can be that you prove that you are over 18 years old without revealing your exact date of birth. Another example could be that you live in one of a range of countries without revealing exactly which country you live in.\n\nYou can read more on\ndeveloper.concordium.software',
         },
         headers: {
             reveal: 'Information to reveal',
@@ -32,8 +33,18 @@ export default {
             missingAttribute: 'The attribute cannot be found on the identity "{{identityName}}"',
         },
     },
+    select: {
+        verifiableCredential: 'Select verifiable credential',
+        accountCredential: 'Select Account',
+    },
+    descriptions: {
+        verifiableCredential: 'Select a verifiable credential to reveal/prove the requested information.',
+        accountCredential:
+            'Select an account associated with the identity whose credentials will be used to reveal/prove the requested information.',
+        unableToProve: 'One or more attributes do not meet the requirements from the verifier.',
+        noCredentialsForThatIssuer:
+            'You do not hold any active credentials from the issuer that the verifier requested.',
+    },
     failedProof: 'Unable to create proof',
     failedProofReason: 'Unable to create proof due to: {{ reason }}',
-    unableToProve:
-        ' {{ dappName }} has requested a proof of identity from you, however you are unable to fulfill the request',
 };